requirements.txt
# Dependencias para MWAA
apache-airflow-providers-amazon==8.12.0
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-http==4.7.0

# Análisis de datos
pandas==2.1.4
numpy==1.24.3
requests==2.31.0

# Análisis de sentimiento
textblob==0.17.1
nltk==3.8.1

# AWS
boto3==1.34.0
botocore==1.34.0

# dbt
dbt-core==1.7.3
dbt-redshift==1.7.3

# Utilidades
python-dotenv==1.0.0

# Core Data Processing
pydantic==2.5.2

# Streaming and Real-time Processing
kinesis-python==0.1.0
confluent-kafka==2.3.0

# Natural Language Processing
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2

# Database and Data Warehouse
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
httpx==0.25.2

# Data Quality and Testing
great-expectations==0.18.19
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Jupyter for Development
jupyter==1.0.0
ipykernel==6.27.1 